CSV File
Finished writing the script to scrape all the nba data i want this time
around into a CSV file. Figured out that if I use the append mode that
there would be blank lines in between all my data, made the writer
object global and changed it to write mode. Fixed the problem. Now I
just wait on the code to finish executing. eta a couple of hours