Color Mapper Works now
reversed the recoding of the team and position columns so that they are
the actual teams and positions and not numerical categorical
assignments. The bokeh plot also separates (colors) each point by
position now. Next step is to be able to select how to color the plots
and to implement a hover option to show which player it is.